ci: add debian-ports architectures
There is no qemu hppa emulation at all, so we don't bother trying.

As for x32, it's supposed to be run as native arch but only with
kernel command line argument "syscall.x32=y" being set on Debian
kernel >= 3.14.15-1. Since there is no way to manipulate kernel command
line on Travis-CI, it's also listed in the exclusion group.